refactor: rewrite to esm
