+++ .github/workflows/test.yml
@@ -4,15 +4,13 @@ on: [push]
 
 jobs:
   test:
-
     runs-on: ubuntu-latest
-
     steps:
     - uses: actions/checkout@master
-    - name: Use Node.js 10.x
+    - name: Use Node.js 14.x
       uses: actions/setup-node@v2.1.5
       with:
-        version: 10.x
+        version: 14.x
     - name: npm install, build, and test
       run: |
         npm ci
