--- .mocharc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  timeout: 5000,
-}
