+++ package.json
@@ -3,7 +3,8 @@
   "description": "Complete and up-to-date info about every release of Electron",
   "repository": "https://github.com/electron/electron-releases",
   "version": "3.678.0",
-  "main": "index.json",
+  "exports": "./index.json",
+  "type": "module",
   "license": "MIT",
   "devDependencies": {
     "@octokit/rest": "^18.0.7",
@@ -13,10 +14,10 @@
     "flat": "^5.0.2",
     "github-slugger": "^1.3.0",
     "got": "^11.8.0",
-    "hubdown": "^2.6.0",
+    "hubdown": "^2.3.2",
     "json-to-markdown-table": "^1.0.0",
     "make-promises-safe": "^5.1.0",
-    "mdast-util-to-string": "^2.0.0",
+    "mdast-util-to-string": "^3.1.0",
     "mocha": "^8.2.0",
     "parse-link-header": "^1.0.1",
     "platform-utils": "^1.2.0",
@@ -26,11 +27,11 @@
   },
   "scripts": {
     "build": "npm run collect && npm run lite && npm run update-readme",
-    "collect": "node script/collect.js",
-    "lite": "node script/lite.js > lite.json",
-    "update-readme": "node script/update-readme.js",
-    "prettier": "prettier --check \"./**/*.{ts,js}\"",
-    "test": "mocha && prettier --check \"./**/*.{ts,js}\" --write",
+    "collect": "node script/collect.mjs --experimental-json-modules",
+    "lite": "node script/lite.js --experimental-json-modules > lite.json",
+    "update-readme": "node script/update-readme.mjs --experimental-json-modules",
+    "prettier": "prettier --check \"./**/*.mjs\"",
+    "test": "mocha --timeout 5000 --experimental-json-modules && prettier --check \"./**/*.mjs\" --write",
     "prepack": "check-for-leaks",
     "prepush": "check-for-leaks",
     "release": "./script/release.sh"
